fix(native-e2e): extract getModelFromEnv to fix circulars
